Update voice-recorder-card.js
Added a notification_id to the persistent notification service. I'm using the size as the notification ID so then in the event data, I can use the size to identify the notification so then I can clear the notification from automation (action: persistent_notification.dismiss).

I don't want to clutter my notifications on my kiosks that I have. Another option would be to have a way to disable notifications for successful recordings, but this works for me.